# api/openapi/openapi.yaml
openapi: 3.0.3
info:
  title: Omnia Dashboard API
  description: REST API for the Omnia AI Agent Orchestration Platform dashboard
  version: 1.0.0
  license:
    name: Apache 2.0
    url: https://www.apache.org/licenses/LICENSE-2.0
servers:
  - url: http://localhost:8082
    description: Local development
  - url: http://omnia-controller-manager.omnia-system.svc.cluster.local:8082
    description: In-cluster
...
paths:
...
  /api/v1/providers:
    get:
      tags: [providers]
      summary: List all Providers
      operationId: listProviders
      parameters:
        - name: namespace
          in: query
          description: Filter by namespace
          schema:
            type: string
      responses:
        '200':
          description: List of providers
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Provider'
        '500':
          $ref: '#/components/responses/InternalError'
...
components:
  responses:
...
    InternalError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
  schemas:
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: string
...
    ObjectMeta:
      type: object
      properties:
        name:
          type: string
        namespace:
          type: string
        uid:
          type: string
        resourceVersion:
          type: string
        creationTimestamp:
          type: string
          format: date-time
        labels:
          type: object
          additionalProperties:
            type: string
        annotations:
          type: object
          additionalProperties:
            type: string
...
    Condition:
      type: object
      properties:
        type:
          type: string
        status:
          type: string
          enum: ['True', 'False', 'Unknown']
        lastTransitionTime:
          type: string
          format: date-time
        reason:
          type: string
        message:
          type: string
...
    Provider:
      type: object
      properties:
        apiVersion:
          type: string
        kind:
          type: string
        metadata:
          $ref: '#/components/schemas/ObjectMeta'
        spec:
          $ref: '#/components/schemas/ProviderSpec'
        status:
          $ref: '#/components/schemas/ProviderStatus'
    ProviderSpec:
      type: object
      properties:
        type:
          type: string
          enum: [claude, openai, gemini, ollama, mock]
        model:
          type: string
        baseURL:
          type: string
          description: Override the provider's default API endpoint
        secretRef:
          $ref: '#/components/schemas/SecretKeyRef'
        defaults:
          $ref: '#/components/schemas/ProviderDefaults'
        pricing:
          $ref: '#/components/schemas/ProviderPricing'
        validateCredentials:
          type: boolean
          description: Enable credential validation on reconciliation
    SecretKeyRef:
      type: object
      required: [name]
      properties:
        name:
          type: string
        key:
          type: string
    ProviderDefaults:
      type: object
      properties:
        temperature:
          type: string
        topP:
          type: string
        maxTokens:
          type: integer
        contextWindow:
          type: integer
        truncationStrategy:
          type: string
          enum: [sliding, summarize, custom]
    ProviderPricing:
      type: object
      properties:
        inputCostPer1K:
          type: string
        outputCostPer1K:
          type: string
        cachedCostPer1K:
          type: string
    ProviderStatus:
      type: object
      properties:
        phase:
          type: string
          enum: [Ready, Error]
        conditions:
          type: array
          items:
            $ref: '#/components/schemas/Condition'
        lastValidatedAt:
          type: string
          format: date-time
        observedGeneration:
          type: integer
...